+++ spec/onlyoffice_webdriver_wrapper/web_driver_click_methods/web_driver_click_on_displayed_spec.rb
@@ -6,8 +6,7 @@ describe OnlyofficeWebdriverWrapper::WebDriver, '#click_on_displayed' do
   let(:webdriver) { described_class.new(:chrome) }
 
   before do
-    webdriver.open("file://#{Dir.pwd}/spec/html_examples/"\
-                   'element_appear_after_click.html')
+    webdriver.open(local_file('element_appear_after_click.html'))
   end
 
   after { webdriver.quit }
@@ -23,8 +22,7 @@ describe OnlyofficeWebdriverWrapper::WebDriver, '#click_on_displayed' do
   end
 
   it 'click_on_displayed correct work if displayed second element' do
-    webdriver.open("file://#{Dir.pwd}/spec/html_examples/"\
-                   'element_appear_after_click_click_on_displayed.html')
+    webdriver.open(local_file('element_appear_after_click_click_on_displayed.html'))
     webdriver.click_on_displayed('//button')
     expect(webdriver.get_element('//*[@id="newElement"]')).not_to be_nil
   end