+++ spec/onlyoffice_webdriver_wrapper/web_driver_set_style_show_by_xpath_spec.rb
@@ -6,8 +6,7 @@ describe OnlyofficeWebdriverWrapper::WebDriver, '#set_style_show_by_xpath' do
   let(:webdriver) { described_class.new(:chrome) }
 
   before do
-    webdriver.open("file://#{Dir.pwd}/spec/html_examples/"\
-                   'set_style_show_by_xpath.html')
+    webdriver.open(local_file('set_style_show_by_xpath.html'))
   end
 
   after { webdriver.quit }