+++ spec/spec_helper.rb
@@ -8,3 +8,11 @@ if ENV['CI']
   SimpleCov.formatter = SimpleCov::Formatter::CoberturaFormatter
 end
 require 'onlyoffice_webdriver_wrapper'
+
+# Get full path for open in browser of local file
+# @param file_name [String] name of file to load
+# @return [String] path to local file
+def local_file(file_name)
+  "file://#{Dir.pwd}/spec/html_examples/"\
+    "#{file_name}"
+end