+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Patch `Headless` gem to be compatible with `ruby-3.2`
 * Remove `w3school.com` usage from specs
 * Add timeout options to `WebDriver#switch_to_popup` method
+* Add `spec` method to get path of local file
 
 ## 1.6.0 (2022-05-25)